# openapi/cinefield.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cinefield API & Event Contracts",
    "version": "1.0.0",
    "description": "Generated by scripts/generate-openapi.ts — DO NOT EDIT BY HAND. Domain event schemas are derived from src/lib/events/event-schemas.ts (the real runtime validation source of truth). HTTP paths are a hand-verified, partial subset — see that script's header for scope."
  },
  "paths": {
    "/api/health/live": {
      "get": {
        "operationId": "getHealthLive",
        "summary": "Public liveness probe — process is answering, nothing more.",
        "responses": {
          "200": {
            "description": "The process is alive.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "status",
                    "timestamp"
                  ],
                  "properties": {
                    "status": {
                      "type": "string",
                      "const": "HEALTHY"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/privileged-actions/decide": {
      "post": {
        "operationId": "decidePrivilegedAction",
        "summary": "A second, distinct admin approves or rejects a pending Phase 16-E two-person privileged action.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "requestId",
                  "action",
                  "targetType",
                  "decision"
                ],
                "properties": {
                  "requestId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "action": {
                    "type": "string",
                    "pattern": "^[a-z][a-z0-9_.]{1,80}$"
                  },
                  "targetType": {
                    "type": "string",
                    "pattern": "^[a-z][a-z0-9_]{1,64}$"
                  },
                  "targetId": {
                    "type": [
                      "string",
                      "null"
                    ]
                  },
                  "decision": {
                    "type": "string",
                    "enum": [
                      "approve",
                      "reject"
                    ]
                  },
                  "reasonCode": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "pattern": "^[a-z][a-z0-9_]{1,64}$"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The decision outcome — never the underlying action's execution result.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "outcome"
                  ],
                  "properties": {
                    "outcome": {
                      "type": "string",
                      "enum": [
                        "INVALID_REQUEST",
                        "UNKNOWN_ACTION",
                        "ROLE_NOT_PERMITTED",
                        "STEP_UP_REQUIRED",
                        "SELF_APPROVAL_BLOCKED",
                        "NO_MATCHING_REQUEST",
                        "REQUEST_EXPIRED",
                        "REJECTED",
                        "APPROVAL_RECORDED",
                        "SOURCE_UNAVAILABLE"
                      ]
                    },
                    "reasonCode": {
                      "type": "string"
                    },
                    "approvals": {
                      "type": "integer",
                      "minimum": 0
                    },
                    "required": {
                      "type": "integer",
                      "minimum": 0
                    },
                    "satisfied": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Malformed request body."
          },
          "404": {
            "description": "Caller is not a console admin (opaque, matches every other Phase 16 admin route)."
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "DomainEventEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "eventId",
          "eventType",
          "eventVersion",
          "aggregateType",
          "aggregateId",
          "occurredAt",
          "payload"
        ],
        "properties": {
          "eventId": {
            "type": "string",
            "format": "uuid"
          },
          "eventType": {
            "type": "string",
            "pattern": "^[a-z][a-z0-9-]*\\.[a-z][a-z0-9-]*$"
          },
          "eventVersion": {
            "type": "integer",
            "minimum": 1
          },
          "aggregateType": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "aggregateId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "occurredAt": {
            "type": "string",
            "format": "date-time"
          },
          "traceId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "tenantId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255
          },
          "payload": {
            "type": "object",
            "additionalProperties": false
          }
        }
      },
      "AssetRejectedPayloadV1": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "assetId"
        ],
        "properties": {
          "assetId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "AssetReleasedPayloadV1": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "assetId"
        ],
        "properties": {
          "assetId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "AuditFlagChangedPayloadV1": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "flagKey",
          "riskTier",
          "newValue"
        ],
        "properties": {
          "flagKey": {
            "type": "string",
            "pattern": "^[a-z][a-z0-9_.]{1,80}$"
          },
          "riskTier": {
            "type": "string",
            "enum": [
              "OPERATOR",
              "HIGH_RISK_TIER0"
            ]
          },
          "newValue": {
            "type": "string",
            "maxLength": 500
          },
          "previousValue": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 500
          }
        }
      },
      "CreditRefundedPayloadV1": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "reservationId",
          "returned"
        ],
        "properties": {
          "reservationId": {
            "type": "string",
            "format": "uuid"
          },
          "generationId": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 64
          },
          "returned": {
            "type": "integer",
            "minimum": 0
          },
          "reason": {
            "type": "string",
            "pattern": "^[a-z][a-z0-9_]{1,64}$"
          }
        }
      },
      "CreditReservedPayloadV1": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "reservationId",
          "amount"
        ],
        "properties": {
          "reservationId": {
            "type": "string",
            "format": "uuid"
          },
          "generationId": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 64
          },
          "amount": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "CreditSettledPayloadV1": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "reservationId",
          "charged"
        ],
        "properties": {
          "reservationId": {
            "type": "string",
            "format": "uuid"
          },
          "generationId": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 64
          },
          "charged": {
            "type": "integer",
            "minimum": 0
          },
          "returned": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "GenerationCancelledPayloadV1": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "generationId"
        ],
        "properties": {
          "generationId": {
            "type": "string",
            "format": "uuid"
          },
          "provider": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          }
        }
      },
      "GenerationCompletedPayloadV1": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "generationId",
          "provider"
        ],
        "properties": {
          "generationId": {
            "type": "string",
            "format": "uuid"
          },
          "provider": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "attemptNo": {
            "type": "integer",
            "minimum": 1
          },
          "durationMs": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "GenerationCreatedPayloadV1": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "generationId",
          "generationType",
          "provider",
          "model"
        ],
        "properties": {
          "generationId": {
            "type": "string",
            "format": "uuid"
          },
          "generationType": {
            "type": "string",
            "enum": [
              "image",
              "video",
              "audio",
              "text"
            ]
          },
          "provider": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "model": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          }
        }
      },
      "GenerationFailedPayloadV1": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "generationId",
          "provider",
          "errorCode"
        ],
        "properties": {
          "generationId": {
            "type": "string",
            "format": "uuid"
          },
          "provider": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "errorCode": {
            "type": "string",
            "pattern": "^[A-Z][A-Z0-9_]{1,64}$"
          },
          "attemptNo": {
            "type": "integer",
            "minimum": 1
          }
        }
      },
      "GenerationProcessingPayloadV1": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "generationId"
        ],
        "properties": {
          "generationId": {
            "type": "string",
            "format": "uuid"
          },
          "provider": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          }
        }
      },
      "ProviderFailedPayloadV1": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "generationId",
          "attemptId",
          "provider",
          "errorCode"
        ],
        "properties": {
          "generationId": {
            "type": "string",
            "format": "uuid"
          },
          "attemptId": {
            "type": "string",
            "format": "uuid"
          },
          "provider": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "errorCode": {
            "type": "string",
            "pattern": "^[A-Z][A-Z0-9_]{1,64}$"
          },
          "submissionEvidence": {
            "type": "string",
            "enum": [
              "none",
              "ambiguous",
              "job"
            ]
          }
        }
      },
      "ProviderSubmittedPayloadV1": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "generationId",
          "attemptId",
          "provider",
          "submissionEvidence"
        ],
        "properties": {
          "generationId": {
            "type": "string",
            "format": "uuid"
          },
          "attemptId": {
            "type": "string",
            "format": "uuid"
          },
          "provider": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "providerJobId": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 500
          },
          "submissionEvidence": {
            "type": "string",
            "enum": [
              "none",
              "ambiguous",
              "job"
            ]
          }
        }
      },
      "SecurityWarningPayloadV1": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "reasonCode",
          "severity"
        ],
        "properties": {
          "reasonCode": {
            "type": "string",
            "pattern": "^[a-z][a-z0-9_]{1,64}$"
          },
          "severity": {
            "type": "string",
            "enum": [
              "info",
              "low",
              "medium",
              "high"
            ]
          }
        }
      }
    }
  }
}